+++ scripts/dumpRobloxTypes.py
@@ -349,8 +349,7 @@ EXTRA_MEMBERS = {
 # Hardcoded types
 # These will go before anything else, and are useful to define for other tools
 START_BASE = """
-type Content = string
-type ContentId = any
+type ContentId = string
 type ProtectedString = string
 type BinaryString = string
 type QDir = string