+++ CHANGELOG.md
@@ -6,9 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.40.0] - 2025-03-01
+
+### Added
+
+- Added configuration `luau-lsp.types.disabledGlobals` to support removing globals from the main scope for analysis.
+  Accepts a list of libraries or library methods (e.g., `table`, `string.split`,
+  etc.) ([#888](https://github.com/JohnnyMorganz/luau-lsp/issues/888))
+
 ### Changed
 
 - Sync to upstream Luau 0.663
+- Improved error reporting when the studio plugin server in VSCode fails to start up (typically due to port already in use) ([#936](https://github.com/JohnnyMorganz/luau-lsp/issues/936))
 
 ### Fixed